refactor : move validation logic to validate package, remove duplicate beef validation
